pyproject.toml
[tool.poetry]
name = "goit-web-hw-07"
version = "0.1.0"
description = ""
authors = ["R.Anatolii <tv.weplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
faker = "^30.1.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
